zdtarif_bot/requirements.txt
pandas
python-telegram-bot
python-dotenv